+++ maia/algo/part/geometry/__init__.py
@@ -1,5 +1,7 @@
 from .centers import compute_cell_center,\
                      compute_edge_center,\
                      compute_face_center,\
-                     _compute_zone_centers,\
-                     compute_zone_centers
+                     _compute_elements_center,\
+                     compute_elements_center
+
+from .measures import _compute_elements_measure, compute_elements_measure